# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VTEI Student API",
        "description": "REST API для мобільного додатку студентів ВТЕІ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/faq/{group_id}/questions": {
            "get": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Питання групи FAQ",
                "description": "Повертає всі питання і відповіді вказаної групи.",
                "operationId": "e7890485333e8fcf596f2f2a733a23f6",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список питань"
                    },
                    "404": {
                        "description": "Групу не знайдено"
                    }
                }
            }
        },
        "/faq/question": {
            "post": {
                "tags": [
                    "FAQ"
                ],
                "summary": "Задати питання",
                "description": "Додає нове питання у вказану групу FAQ.",
                "operationId": "de3e532f81aada881da2eb098c1dc478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question",
                                    "group_id"
                                ],
                                "properties": {
                                    "question": {
                                        "description": "Текст питання",
                                        "type": "string"
                                    },
                                    "group_id": {
                                        "description": "ID групи",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Питання додано"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                }
            }
        },
        "/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Успішність по семестрах",
                "description": "Повертає модульні оцінки та результати екзаменів/заліків, згруповані по навчальних роках і семестрах.",
                "operationId": "219f2a64ee0e5767a4aa4f1c65debdbc",
                "responses": {
                    "200": {
                        "description": "Успішність згрупована по роках і семестрах"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/grades/modules": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Модульні картки поточного року",
                "description": "Повертає модульні картки студента за поточний навчальний рік.",
                "operationId": "48c6a2380d5b0d537caef8b27dddbb0c",
                "responses": {
                    "200": {
                        "description": "Список карток"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/grades/modules/{card_id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Деталі модульної картки",
                "description": "Повертає розбивку балів по окремих завданнях кожного модуля.",
                "operationId": "b86102ece1d50050cd892de62773282f",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деталі картки"
                    },
                    "404": {
                        "description": "Не знайдено"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/grades/markbook": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Залікова книжка",
                "description": "Всі результати екзаменів та заліків, згруповані по навчальних роках і семестрах.",
                "operationId": "d7ba528efff244324e907793071a4fa7",
                "responses": {
                    "200": {
                        "description": "Залікова книжка"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}